requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
algoliasearch==3.0.0
amqp==5.2.0
anyio==4.1.0
asgiref==3.7.2
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
bandit==1.7.5
billiard==4.2.0
black==23.11.0
celery==5.3.6
certifi==2023.7.22
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
dill==0.3.7
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
flake8==6.1.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httptools==0.6.1
idna==3.4
inflection==0.5.1
infobip-api-python-sdk==5.0.1
isort==5.12.0
kombu==5.3.4
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
pbr==6.0.0
phonenumbers==8.13.24
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==1.10.13
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2022.3.15
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
tomlkit==0.12.2
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
whitenoise==6.6.0
yarl==1.9.2
